+++ util/util.go
@@ -1,4 +1,4 @@
-package indexer
+package util
 
 import (
 	"crypto/rand"
@@ -6,8 +6,7 @@ import (
 	"math/big"
 )
 
-// GenerateSalts generates `numKeys` salts with length `lenSalt` for the usage
-// of `SecureIndexBuilder`.
+// GenerateSalts generates `numKeys` salts with length `lenSalt`.
 func GenerateSalts(numKeys int, lenSalt uint) (salts [][]byte) {
 	if lenSalt < 8 {
 		fmt.Println("Error in generating the salts: lenSalt must be at leat 8")
@@ -24,9 +23,9 @@ func GenerateSalts(numKeys int, lenSalt uint) (salts [][]byte) {
 	return
 }
 
-// Returns a random 64-bit unsigned integer in the range of [0, n).  Panics if
-// n <= 0.
-func randUint64n(n uint64) uint64 {
+// RandUint64n returns a random 64-bit unsigned integer in the range of [0, n).
+// Panics if n <= 0.
+func RandUint64n(n uint64) uint64 {
 	i := new(big.Int)
 	i.SetUint64(n)
 	num, _ := rand.Int(rand.Reader, i)
